requirements.txt
# Required modules
dotenv
halo
logging
pychalk
requests
